+++ WeeClik/Extensions/Extension+Notification.swift
@@ -1,0 +1,13 @@
+//
+//  Extension+Notification.swift
+//  WeeClik
+//
+//  Created by Herrick Wolber on 23/11/2019.
+//  Copyright © 2019 Herrick Wolber. All rights reserved.
+//
+
+import Foundation
+
+extension Notification.Name {
+    static let didSendGroupeFavorisSMS = Notification.Name("didSendGroupeFavorisSMS")
+}