+++ WeeClik/HelperAndKeys.swift
@@ -10,76 +10,10 @@ import Foundation
 import UIKit
 import Parse
 import MapKit
-import MessageUI
 import CRNotifications
 import SwiftDate
 
 class HelperAndKeys {
-
-    static func showAlertWithMessage(theMessage: String, title: String, viewController: UIViewController) {
-        let alertViewController = UIAlertController.init(title: title, message: theMessage, preferredStyle: UIAlertController.Style.alert)
-        let defaultAction = UIAlertAction.init(title: "OK".localized(), style: .cancel) { (_) -> Void in
-            alertViewController.dismiss(animated: true, completion: nil)
-        }
-        alertViewController.addAction(defaultAction)
-        viewController.present(alertViewController, animated: true, completion: nil)
-    }
-
-    static func showSettingsAlert(withTitle title: String, withMessage message: String, presentFrom viewController: UIViewController) {
-        let alertController = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        let settingsAction = UIAlertAction(title: "Réglages".localized(), style: .default) { (_) -> Void in
-            guard let settingsUrl = URL(string: UIApplication.openSettingsURLString) else {return}
-            if UIApplication.shared.canOpenURL(settingsUrl) {UIApplication.shared.open(settingsUrl, completionHandler: nil)}
-        }
-        let cancelAction = UIAlertAction(title: "Annuler".localized(), style: .default, handler: nil)
-        alertController.addAction(cancelAction)
-        alertController.addAction(settingsAction)
-        viewController.present(alertController, animated: true, completion: nil)
-    }
-    
-    static func visitWebsite(site: String, controller: UIViewController) {
-
-        let alertViewController = UIAlertController.init(title: "Sortir de l'application ?".localized(), message: "Vous allez être redirigé vers le site web du commerçant.\n Et ainsi quitter l'application Weeclik.\n Voulez vous continuer ?".localized(), preferredStyle: UIAlertController.Style.alert)
-        let defaultAction = UIAlertAction.init(title: "OK".localized(), style: UIAlertAction.Style.default) { (_) -> Void in
-            if let url = URL(string: site), UIApplication.shared.canOpenURL(url) {
-                UIApplication.shared.open(url)
-            }
-        }
-        let cancelAction = UIAlertAction.init(title: "Annuler".localized(), style: UIAlertAction.Style.destructive) {(_) -> Void in}
-        alertViewController.addAction(cancelAction)
-        alertViewController.addAction(defaultAction)
-        controller.present(alertViewController, animated: true, completion: nil)
-    }
-
-    static func sendFeedBackOrMessageViaMail(messageToSend: String, isFeedBackMsg: Bool, commerceMail: String, controller: UIViewController) {
-        let messageAdded: String
-        let versionNumber = Bundle.main.object(forInfoDictionaryKey: "CFBundleShortVersionString") as! String
-
-        if !isFeedBackMsg {
-            messageAdded = "<br><br>Envoyé depuis l'application iOS Weeclik.<br><br>Téléchargez-la ici : http://www.google.fr/".localized()
-        } else {
-            messageAdded = "<br><br>Envoyé depuis l'application iOS Weeclik.<br><br>Numéro de version de l'app : \(versionNumber)".localized()
-        }
-//                let allowedCharacters = NSCharacterSet.urlFragmentAllowed
-        let finalMessage = messageToSend.appending(messageAdded)
-
-        if MFMailComposeViewController.canSendMail() {
-            let composeVC = MFMailComposeViewController()
-
-            // Configure the fields of the interface.
-            composeVC.setSubject("Demande de contact via Weeclik".localized())
-            composeVC.setToRecipients([commerceMail])
-            composeVC.setMessageBody(finalMessage, isHTML: true)
-
-            composeVC.navigationBar.barTintColor = UIColor.white
-
-            // Present the view controller modally.
-            controller.present(composeVC, animated: true, completion: nil)
-        } else {
-            showAlertWithMessage(theMessage: "Il semblerait que vous n'ayez pas configuré votre boîte mail depuis votre téléphone.".localized(), title: "Erreur".localized(), viewController: controller)
-        }
-
-    }
 
     static func setUserDefaultsValue(value: Any, forKey key: String) {
         let standardUserDefaults = UserDefaults.standard
@@ -222,7 +156,6 @@ class HelperAndKeys {
             return true
         }
         let isAfterIntervalle = Date().isAfterDate(date + 7.days, granularity: .second)
-        print("isAfterIntervalle : \(isAfterIntervalle)")
         if isAfterIntervalle {
             removeCommerce(forCommerceId: objectId)
             return true
@@ -279,40 +212,57 @@ class HelperAndKeys {
     // TODO: Construire une veritable requette de stats (perfection)
     static func saveStatsInDb(commerce: PFObject, user: PFUser? = nil) {
         // Il y a un compte utilisateur, on met donc a jour ses stats
-        if let utilisateur = user {
+        if let user = user {
             let query = PFQuery(className: "StatsPartage")
             query.whereKey("commercePartage", equalTo: commerce)
-            query.whereKey("utilisateurPartageur", equalTo: utilisateur)
+            query.whereKey("utilisateurPartageur", equalTo: user)
             query.includeKey("commercePartage")
             query.getFirstObjectInBackground { (object, error) in
                 if let error = error {
                     print(error.desc)
-                    if error.localizedDescription == "No results matched the query." {
+                    if error.code == 101 {
                         // Aucun objet trouvé on en crée un
                         let parseObj = PFObject(className: "StatsPartage")
                         parseObj["commercePartage"] = commerce
-                        parseObj["utilisateurPartageur"] = utilisateur
+                        parseObj["utilisateurPartageur"] = user
                         parseObj["nbrPartage"] = 1
                         parseObj["mes_partages_dates"] = [Date()]
-                        parseObj.saveInBackground()
+                        parseObj.saveInBackground { (_ success: Bool, error) in
+                            if let _ = error {
+                                
+                            } else {
+                                // On met a jour les stats du commerce sans utilisateur
+                                let commerceQuery = PFQuery(className: "Commerce")
+                                commerceQuery.getObjectInBackground(withId: commerce.objectId!) { (comm, error) in
+                                    if let error = error {
+                                        ParseErrorCodeHandler.handleUnknownError(error: error)
+                                    } else if let comm = comm {
+                                        comm.incrementKey("nombrePartages")
+                                        comm.saveInBackground()
+                                    }
+                                }
+                            }
+                        }
                     } else {
                         print("Error in save stat partage func - HelperAndKeys - saveStatsInDb")
                         ParseErrorCodeHandler.handleUnknownError(error: error)
                     }
 
-                } else if let object = object {
-                    let sharingObject = object
+                } else if let sharingObject = object {
                     sharingObject.incrementKey("nbrPartage")
                     sharingObject.add(Date(), forKey: "mes_partages_dates")
-                    let theCommerce = sharingObject["commercePartage"] as! PFObject
-                    theCommerce.incrementKey("nombrePartages")
-                    PFObject.saveAll(inBackground: [sharingObject, theCommerce])
+                    if let theCommerce = sharingObject["commercePartage"] as? PFObject {
+                        theCommerce.incrementKey("nombrePartages")
+                        PFObject.saveAll(inBackground: [sharingObject, theCommerce])
+                    } else {
+                        sharingObject.saveInBackground()
+                    }
                 }
             }
         } else {
             // On met a jour les stats du commerce sans utilisateur
             let commerceQuery = PFQuery(className: "Commerce")
-            commerceQuery.getObjectInBackground(withId: commerce.objectId!.description) { (comm, error) in
+            commerceQuery.getObjectInBackground(withId: commerce.objectId!) { (comm, error) in
                 if let error = error {
                     ParseErrorCodeHandler.handleUnknownError(error: error)
                 } else if let comm = comm {