+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Add
 - Video compression for a faster upload to database.
 
+## [3.3.5] - 2019-11-24
+### Add
+- Add precheck for app store in beta & screenshot lanes.
+
+### Changed
+- Sharing icon was not called for SMS & Whatsapp sharing.
+- Method function for globally showing alerts in App.
+- Sometimes sharing function would not increase number information of commerce.
+- Groupe sharing now trigger the already send check in commerce detail.
+- SwiftLint to don't include Extensions folder in build check
+
 ## [3.3.4] - 2019-11-17
 ### Changed
 - Set minimum iOS version from 13 to 11.